Moved all compiler unit tests to cgreen

+++ compiler/ifid.c
@@ -108,15 +108,15 @@ static Bool isValidUUID(char *uuid)
 
   if (strlen(uuid) != 45) return FALSE;
   if (strncmp(uuid, "UUID://", 7) != 0) return FALSE;
-  for (i = 7; i <= 14; i++) if (!isxdigit(uuid[i])) return FALSE;
+  for (i = 7; i <= 14; i++) if (!isxdigit((uint)uuid[i])) return FALSE;
   if (uuid[15] != '-') return FALSE;
-  for (i = 16; i <= 19; i++) if (!isxdigit(uuid[i])) return FALSE;
+  for (i = 16; i <= 19; i++) if (!isxdigit((uint)uuid[i])) return FALSE;
   if (uuid[20] != '-') return FALSE;
-  for (i = 21; i <= 24; i++) if (!isxdigit(uuid[i])) return FALSE;
+  for (i = 21; i <= 24; i++) if (!isxdigit((uint)uuid[i])) return FALSE;
   if (uuid[25] != '-') return FALSE;
-  for (i = 26; i <= 29; i++) if (!isxdigit(uuid[i])) return FALSE;
+  for (i = 26; i <= 29; i++) if (!isxdigit((uint)uuid[i])) return FALSE;
   if (uuid[30] != '-') return FALSE;
-  for (i = 31; i <= 42; i++) if (!isxdigit(uuid[i])) return FALSE;
+  for (i = 31; i <= 42; i++) if (!isxdigit((uint)uuid[i])) return FALSE;
   if (strcmp(&uuid[43], "//") != 0) return FALSE;
   return TRUE;
 }
